requirements.txt
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
openai>=1.60.0
faiss-cpu>=1.8.0
pydantic>=2.8.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.1
pandas>=2.0.0
streamlit>=1.40.0
tiktoken>=0.7.0
typing_extensions>=4.0.0
pypdf>=3.0.0
langchain-text-splitters>=0.3.0